Pull out shared hook
